feat: 상세내용 가져오기(#27)
