+++ package.json
@@ -15,6 +15,7 @@
     "@tanstack/react-query": "^5.17.9",
     "axios": "^1.6.5",
     "classnames": "^2.5.1",
+    "dayjs": "^1.11.10",
     "framer-motion": "^10.17.9",
     "moment": "^2.30.1",
     "next": "14.0.4",
