+++ src/app/stray-dogs/[desertionNo]/page.module.scss
@@ -1,0 +1,48 @@
+.container {
+  width: 1200px;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  margin: 100px auto;
+}
+
+.explanationWrap {
+  width: 1200px;
+  height: 100%;
+  margin: 50px auto;
+  display: flex;
+  justify-content: space-between;
+  border: 2px solid #0ac4b9;
+  border-radius: 20px;
+  padding: 60px 0;
+}
+
+.imageWrap {
+  width: 800px;
+  height: 500px;
+  margin: 50px auto;
+}
+
+.imageWrap > img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  border: 5px solid white;
+  box-shadow: 0px 0px 8px 0px rgba(0, 0, 0, 0.5);
+}
+
+.titleColumn {
+  width: 30%;
+  font-size: 18px;
+  line-height: 35px;
+  margin: auto 0 auto 100px;
+}
+
+.contentColumn {
+  width: 70%;
+  font-size: 18px;
+  gap: 10px;
+  margin: auto 100px auto 0;
+  line-height: 35px;
+}
