--- src/app/stray-dogs/[id]/page.module.scss
@@ -1,12 +1,0 @@
-.container {
-  width: 1200px;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  margin: auto;
-}
-
-.explanationWrap {
-  display: flex;
-}
