+++ src/app/stray-dogs/page.module.scss
@@ -8,14 +8,43 @@
 }
 
 .filterWrap {
-  width: 100%;
+  width: 1200px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   height: 130px;
-  background-color: antiquewhite;
+  // border-bottom: 2px solid #b0b0b0;
+  border-radius: 20px;
+  box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.3);
   margin: 100px auto 50px auto;
 }
 
+.calender {
+  width: 500px;
+  margin: auto 30px;
+}
+
+.datePicker {
+  width: 150px;
+  height: 30px;
+  border: 1px solid #0ac4b9;
+  border-radius: 5px;
+  padding-left: 5px;
+  margin-right: 50px;
+  text-align: center;
+  cursor: pointer;
+}
+
+.selectCity {
+  width: 150px;
+  height: 30px;
+  border: 1px solid #0ac4b9;
+  border-radius: 5px;
+  margin: auto 30px;
+}
+
 .gridContainer {
-  width: 60vw;
+  width: 1200px;
   height: 100%;
   display: grid;
   justify-content: space-between;
@@ -33,7 +62,7 @@
 .listCard {
   width: 100%;
   height: 430px;
-  border: 1px solid #b0b0b0;
+  // border: 1px solid #b0b0b0;
   box-shadow: 0px 0px 5px 0px rgba(0, 0, 0, 0.3);
   overflow: hidden;
   border-radius: 10px;
