+++ src/app/stray-dogs/page.tsx
@@ -4,14 +4,20 @@ import { getStrayList } from '@/apis/stray';
 import style from './page.module.scss';
 import { useQuery } from '@tanstack/react-query';
 import Image from 'next/image';
+import DatePicker from 'react-datepicker';
+import 'react-datepicker/dist/react-datepicker.css';
 import { FaMapMarkerAlt } from 'react-icons/fa';
 import { FaCalendarDays } from 'react-icons/fa6';
 import { PiGenderIntersexFill } from 'react-icons/pi';
 import { FaDog } from 'react-icons/fa6';
 import Loading from '../_components/layout/loading/Loading';
 import Link from 'next/link';
+import { useState } from 'react';
+import regionList from '../../data/regionList.json';
 
 const StrayDogs = () => {
+  const [startDate, setStartDate] = useState<Date | null>(new Date());
+  const [endDate, setEndDate] = useState<Date | null>(new Date());
   const {
     isLoading,
     isError,
@@ -40,7 +46,42 @@ const StrayDogs = () => {
 
   return (
     <div className={style.container}>
-      <div className={style.filterWrap}>필터</div>
+      <div className={style.filterWrap}>
+        <p>기간</p>
+        <div className={style.calender}>
+          <DatePicker
+            className={style.datePicker}
+            dateFormat="yyyy-MM-dd"
+            shouldCloseOnSelect // 날짜를 선택하면 datepicker가 자동으로 닫힘
+            minDate={new Date('2023-01-01')} // minDate 이전 날짜 선택 불가
+            maxDate={new Date()} // maxDate 이후 날짜 선택 불가
+            selected={startDate}
+            startDate={startDate}
+            endDate={endDate}
+            onChange={(date) => setStartDate(date)}
+          />
+          <DatePicker
+            className={style.datePicker}
+            dateFormat="yyyy-MM-dd"
+            shouldCloseOnSelect
+            selected={endDate}
+            startDate={startDate}
+            endDate={endDate}
+            minDate={startDate}
+            maxDate={new Date()} // maxDate 이후 날짜 선택 불가
+            onChange={(date) => setEndDate(date)}
+          />
+        </div>
+        <p>지역</p>
+        <select className={style.selectCity}>
+          {regionList.map((region, index) => {
+            return <option key={index}>{region.name}</option>;
+          })}
+        </select>
+        <select className={style.selectCity}>
+          <option>군/구</option>;
+        </select>
+      </div>
       <div className={style.gridContainer}>
         {strayList?.map((list, index) => {
           const formatNoticeEdt = formatDate(list.noticeEdt);
